requirements.txt
Flask==2.3.2
tensorflow==2.13.0
gunicorn==21.2.0
Pillow==10.0.0
Werkzeug==2.3.6
numpy==1.24.4
